Create styles.css (version 1)
Not a lot of effort so far was put into styling this project seeing this was just used to practice Javascript code. I may come back to it at some point. Things I would change are preventing the table borders from expanding to the boarders of the webpage; color pallet; stacking two input fields on top of eachother and floating their corresponding titles on the left hand side; increasing input field lengths...